Refactor: put gameState related interfaces to game.ts + move pgn parser to gameState to bff folder
